+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Added
+- Adds unit tests to passwordless recipe 
+
+
 ### Fixes
 - Fixes existing action to run go mod tidy in the examples folder
 
@@ -19,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adds more unit tests for thirdpary email password recipe
 - Adds test to jwt recipe
 - Adds test to opendID recipe
+
 
 ### Fixes
 - Third party sign in up API response correction.